exp_01/requirements.txt
transformers
torch
pytorch-lightning
hydra-core==1.0.4
wandb
torchmetrics
datasets
nltk
pandas